feat:created locationcontroller
